+++ Monofoxe/Monofoxe.Engine/Drawing/RectangleShape.cs
@@ -3,33 +3,8 @@ using Microsoft.Xna.Framework.Graphics;
 
 namespace Monofoxe.Engine.Drawing
 {
-	public class RectangleShape : Drawable
-	{
-		public Vector2 Size = Vector2.One;
-
-		/// <summary>
-		/// If false, circle will be filled with solid color. If true, only outline will be drawn.
-		/// </summary>
-		public bool IsOutline = false;
-
-		public Color Color = Color.White;
-
-		public float ZDepth = 0;
-
-		/// <summary>
-		/// Draws a rectangle using instance properties.
-		/// </summary>
-		public override void Draw()
-		{
-			Draw(
-				Position - Size / 2, Position + Size / 2, 
-				IsOutline, 
-				Color, Color, Color, Color, 
-				ZDepth
-			);
-		}
-		
-		
+	public static class RectangleShape
+	{		
 		private static VertexPositionColorTexture[] _rectangleVertices = new VertexPositionColorTexture[4];
 
 		private static short[] _filledRectangleIndices = { 0, 1, 3, 1, 2, 3 }; 
@@ -39,8 +14,8 @@ namespace Monofoxe.Engine.Drawing
 		/// <summary>
 		/// Draws a rectangle using top left and bottom right point.
 		/// </summary>
-		public static void Draw(Vector2 p1, Vector2 p2, bool isOutline) =>
-			Draw(p1, p2, isOutline, GraphicsMgr.CurrentColor, GraphicsMgr.CurrentColor, GraphicsMgr.CurrentColor, GraphicsMgr.CurrentColor);
+		public static void Draw(Vector2 p1, Vector2 p2, ShapeFill fill) =>
+			Draw(p1, p2, fill, GraphicsMgr.CurrentColor, GraphicsMgr.CurrentColor, GraphicsMgr.CurrentColor, GraphicsMgr.CurrentColor);
 
 		/// <summary>
 		/// Draws a rectangle using top left and bottom right point with specified colors for each corner.
@@ -48,7 +23,7 @@ namespace Monofoxe.Engine.Drawing
 		public static void Draw(
 			Vector2 p1, 
 			Vector2 p2, 
-			bool isOutline, 
+			ShapeFill fill, 
 			Color c1, 
 			Color c2, 
 			Color c3, 
@@ -66,7 +41,7 @@ namespace Monofoxe.Engine.Drawing
 			_rectangleVertices[3].Color = c4;
 						
 			GraphicsMgr.VertexBatch.Texture = null;
-			if (isOutline)
+			if (fill == ShapeFill.Outline)
 			{
 				GraphicsMgr.VertexBatch.AddPrimitive(PrimitiveType.LineList, _rectangleVertices, _outlineRectangleIndices);
 			}
@@ -80,16 +55,14 @@ namespace Monofoxe.Engine.Drawing
 		/// <summary>
 		/// Draws a rectangle using center point and size.
 		/// </summary>
-		public static void DrawBySize(Vector2 p, Vector2 size, bool isOutline) =>
-			Draw(p - size / 2, p + size / 2f, isOutline, GraphicsMgr.CurrentColor, GraphicsMgr.CurrentColor, GraphicsMgr.CurrentColor, GraphicsMgr.CurrentColor);
+		public static void DrawBySize(Vector2 p, Vector2 size, ShapeFill fill) =>
+			Draw(p - size / 2, p + size / 2f, fill, GraphicsMgr.CurrentColor, GraphicsMgr.CurrentColor, GraphicsMgr.CurrentColor, GraphicsMgr.CurrentColor);
 
 		/// <summary>
 		/// Draws a rectangle using center point and size with specified colors for each corner.
 		/// </summary>
-		public static void DrawBySize(Vector2 p, Vector2 size, bool isOutline, Color c1, Color c2, Color c3, Color c4) =>
-			Draw(p - size / 2f, p + size / 2f, isOutline, c1, c2, c3, c4);
-
-		
+		public static void DrawBySize(Vector2 p, Vector2 size, ShapeFill fill, Color c1, Color c2, Color c3, Color c4) =>
+			Draw(p - size / 2f, p + size / 2f, fill, c1, c2, c3, c4);
 
 	}
 }