+++ Monofoxe/Monofoxe.Engine/Utils/Vector2Damper.cs
@@ -132,12 +132,13 @@ namespace Monofoxe.Engine.Utils
 
 		public Vector2 Update(Vector2 value)
 		{
-			var dt = (float)Time.Time();
-
-			// Estimating velocity.
-			var speed = (value - _oldValue) / dt;
+			float num = (float)Time.Time();
+			var speed = Vector2.Zero;
+			if (num != 0)
+			{
+				speed = (value - _oldValue) / num;
+			}
 			_oldValue = value;
-
 			return Update(value, speed);
 		}