+++ CHANGELOG.md
@@ -13,17 +13,25 @@
 - Added `UnorderedList` collection.
 - Added `GetArea()`, `GetSignedArea()`, `IsClockwise()` methods to `GameMath`.
 - Added new collision system.
+- Added new drawing methods.
 
 ### Changed
+
 - **BREAKING CHANGE:** `ResourceInfoMgr` now accepts wildcards instead of directory names. For example, `ResourceInfoMgr.GetResourcePaths("Graphics/Trees");` should now be replaced with `ResourceInfoMgr.GetResourcePaths("Graphics/Trees/*");`
+- **BREAKING CHANGE:** Renamed `GetSafeNormalize()` to `SafeNormalize()`.
+- **BREAKING CHANGE:** Removed instances of `Width` and `Height` in `Sprite`, `Frame`, `WindowMgr`, `Camera`, `Surface`, and replaced them with `Size`.
+- **BREAKING CHANGE:** Changed boolean `isOutline` to `ShapeFill` enum for shapes.
 
 ### Fixed
 
 - Fixed `AddComponent<>()` not taking generic type into account.
 - `DirectoryResourceBox` now ignores non-xnb files properly.
+- Fixed division by zero in dampers.
 
 ### Removed
+
 - **BREAKING CHANGE:** Removed camera layer filters.
+- **BREAKING CHANGE:** Removed `Drawable` class and non-static shape fields and methods.
 
 ## [v2.2.0] - *17.09.2022*